requirements.txt
# Core ML/AI
numpy>=2.0
pandas>=2.2
scikit-learn>=1.5
matplotlib>=3.9
seaborn>=0.13

# Deep Learning (uncomment as needed)
# tensorflow>=2.18
# torch>=2.5

# Jupyter
ipykernel
jupyter

# Utilities
python-dotenv
tqdm>=4.67

# Data Processing
openpyxl>=3.1
